Structural cleanup, partials.
